Restructuring The Repo Directory
The goal here is to put all old files inside the directory 00_OLD_WORK_ so as to make the repo a bit more clean. No files is deleated. It's just being put inside the OLD_WORK directory. All operations are being neatly tracked by the GIT and GITHUB's version control. There might be further restructuring of the code too if required at a later commit. Hope that helps