Comitting Pesky .DS_Store Files ||-_-||
These files keep bugging me. I need to figure out a way to get rid of these files. I had already these extensions in .gitignore file but somehow it seems that it's NOT working. No issues. Maybe we'll create a pull request for these problems and then get rid of them once and for all